SEQNG-367 Split files
